fix: refine codes
