+++ packages/core_extensions/default_extension_go/extension.go
@@ -3,10 +3,6 @@
 // Licensed under the Apache License, Version 2.0.
 // See the LICENSE file for more information.
 //
-// Note that this is just an example extension written in the GO programming
-// language, so the package name does not equal to the containing directory
-// name. However, it is not common in Go.
-//
 
 package default_extension_go
