+++ tests/ten_runtime/integration/go/access_property_go/BUILD.gn
@@ -17,7 +17,7 @@ ten_package_test_prepare_app("access_property_go_app") {
     "access_property_go_app/property.json",
   ]
 
-  replace_files_after_install_all = [ "access_property_go_app/ten_packages/extension/default_extension_go/default_extension.go" ]
+  replace_files_after_install_all = [ "access_property_go_app/ten_packages/extension/default_extension_go/extension.go" ]
 
   deps = [
     "//core/src/ten_manager",
