+++ tests/ten_runtime/integration/go/access_property_go/access_property_go_app/ten_packages/extension/default_extension_go/extension.go
@@ -1,13 +1,12 @@
+//
 // Copyright © 2024 Agora
 // This file is part of TEN Framework, an open source project.
 // Licensed under the Apache License, Version 2.0, with certain conditions.
 // Refer to https://github.com/TEN-framework/ten_framework/LICENSE for more
 // information.
 //
-// Note that this is just an example extension written in the GO programming
-// language, so the package name does not equal to the containing directory
-// name. However, it is not common in Go.
-package defaultextension
+
+package default_extension_go
 
 import (
 	"ten_framework/ten"
