+++ tests/ten_runtime/integration/go/exit_signal/exit_signal_app/ten_packages/extension/default_extension_go/extension.go
@@ -1,9 +1,12 @@
+//
 // Copyright © 2024 Agora
 // This file is part of TEN Framework, an open source project.
 // Licensed under the Apache License, Version 2.0, with certain conditions.
 // Refer to https://github.com/TEN-framework/ten_framework/LICENSE for more
 // information.
-package defaultextension
+//
+
+package default_extension_go
 
 import (
 	"encoding/json"
