+++ tests/ten_runtime/integration/go/expired_ten_go/BUILD.gn
@@ -17,7 +17,7 @@ ten_package_test_prepare_app("expired_ten_go_app") {
     "expired_ten_go_app/property.json",
   ]
 
-  replace_files_after_install_all = [ "expired_ten_go_app/ten_packages/extension/default_extension_go/default_extension.go" ]
+  replace_files_after_install_all = [ "expired_ten_go_app/ten_packages/extension/default_extension_go/extension.go" ]
 
   if (ten_enable_package_manager) {
     deps = [
