+++ tests/ten_runtime/integration/go/frequently_cgo_call_go/frequently_cgo_call_go_app/ten_packages/extension/default_extension_go/extension.go
@@ -5,10 +5,7 @@
 // Refer to the "LICENSE" file in the root directory for more information.
 //
 
-// Package defaultextension is just an example extension written in the GO
-// programming language, so the package name does not equal to the containing
-// directory name. However, it is not common in Go.
-package defaultextension
+package default_extension_go
 
 import (
 	"fmt"
