+++ tests/ten_runtime/integration/go/three_extension_cmd_go/three_extension_cmd_go_app/ten_packages/extension/default_extension_go/extension.go
@@ -7,7 +7,7 @@
 // Note that this is just an example extension written in the GO programming
 // language, so the package name does not equal to the containing directory
 // name. However, it is not common in Go.
-package defaultextension
+package default_extension_go
 
 import (
 	"encoding/json"
