+++ tests/ten_runtime/integration/go/transfer_pointer_go/transfer_pointer_go_app/ten_packages/extension/default_extension_go/extension.go
@@ -7,7 +7,7 @@
 // Note that this is just an example extension written in the GO programming
 // language, so the package name does not equal to the containing directory
 // name. However, it is not common in Go.
-package defaultextension
+package default_extension_go
 
 import (
 	"fmt"
@@ -31,6 +31,11 @@ func newExtensionB(name string) ten.Extension {
 	return &extensionB{}
 }
 
+type userData struct {
+	num int
+	str string
+}
+
 func (p *extensionA) OnCmd(
 	tenEnv ten.TenEnv,
 	cmd ten.Cmd,
@@ -39,33 +44,30 @@ func (p *extensionA) OnCmd(
 		fmt.Println("extensionA OnCmd")
 
 		cmdB, _ := ten.NewCmd("B")
+
+		strArray := []string{"hello", "world", "ten"}
+		cmdB.SetProperty("array", &strArray)
+
+		propMap := make(map[string]interface{})
+		propMap["paramA"] = "A"
+		cmdB.SetProperty("map", &propMap)
+
+		data := &userData{num: 2, str: "str"}
+		cmdB.SetProperty("struct", data)
+		data.num = 3
+
 		tenEnv.SendCmd(cmdB, func(r ten.TenEnv, cs ten.CmdResult) {
 			detail, err := cs.GetPropertyString("detail")
 			if err != nil {
 				cmdResult, _ := ten.NewCmdResult(ten.StatusCodeError)
 				cmdResult.SetPropertyString("detail", err.Error())
-				tenEnv.ReturnResult(cmdResult, cmd)
-				return
-			}
-
-			if detail != "this is extensionB." {
-				cmdResult, _ := ten.NewCmdResult(ten.StatusCodeError)
-				cmdResult.SetPropertyString("detail", "wrong detail")
-				tenEnv.ReturnResult(cmdResult, cmd)
-				return
-			}
-
-			password, err := cs.GetPropertyString("password")
-			if err != nil {
-				cmdResult, _ := ten.NewCmdResult(ten.StatusCodeError)
-				cmdResult.SetPropertyString("detail", err.Error())
-				tenEnv.ReturnResult(cmdResult, cmd)
+				r.ReturnResult(cmdResult, cmd)
 				return
 			}
 
 			cmdResult, _ := ten.NewCmdResult(ten.StatusCodeOk)
-			cmdResult.SetPropertyString("detail", password)
-			tenEnv.ReturnResult(cmdResult, cmd)
+			cmdResult.SetPropertyString("detail", detail)
+			r.ReturnResult(cmdResult, cmd)
 		})
 	}()
 }
@@ -79,19 +81,44 @@ func (p *extensionB) OnCmd(
 
 		cmdName, _ := cmd.GetName()
 		if cmdName == "B" {
-			statusCmd, err := ten.NewCmdResult(
-				ten.StatusCodeOk,
-			)
+			strArrayPtr, err := cmd.GetPropertyPtr("array")
 			if err != nil {
-				cmdResult, _ := ten.NewCmdResult(ten.StatusCodeError)
-		cmdResult.SetPropertyString("detail", err.Error())
-		tenEnv.ReturnResult(cmdResult, cmd)
-				return
+				panic("failed to get prop: array")
+			}
+			strArray := *(strArrayPtr.(*[]string))
+
+			if len(strArray) != 3 || strArray[0] != "hello" ||
+				strArray[1] != "world" ||
+				strArray[2] != "ten" {
+				panic("should not happen")
 			}
 
-			statusCmd.SetProperty("detail", "this is extensionB.")
-			statusCmd.SetProperty("password", "password")
-			tenEnv.ReturnResult(statusCmd, cmd)
+			mapPtr, err := cmd.GetPropertyPtr(
+				"map",
+			)
+			if err != nil {
+				panic("failed to get prop: map")
+			}
+			m := *(mapPtr.(*map[string]interface{}))
+
+			if m["paramA"] != "A" {
+				panic("should not happen")
+			}
+
+			structPtr, err := cmd.GetPropertyPtr("struct")
+			if err != nil {
+				panic("failed to get prop: struct")
+			}
+
+			structData := structPtr.(*userData)
+
+			if structData.num != 3 || structData.str != "str" {
+				panic("should not happen")
+			}
+
+			cmdResult, _ := ten.NewCmdResult(ten.StatusCodeOk)
+			cmdResult.SetPropertyString("detail", "ten")
+			tenEnv.ReturnResult(cmdResult, cmd)
 		}
 	}()
 }
