+++ tests/ten_runtime/integration/go/two_extension_one_group_cmd_go/BUILD.gn
@@ -17,7 +17,7 @@ ten_package_test_prepare_app("two_extension_one_group_cmd_go_app") {
     "two_extension_one_group_cmd_go_app/property.json",
   ]
 
-  replace_files_after_install_all = [ "two_extension_one_group_cmd_go_app/ten_packages/extension/default_extension_go/default_extension.go" ]
+  replace_files_after_install_all = [ "two_extension_one_group_cmd_go_app/ten_packages/extension/default_extension_go/extension.go" ]
 
   if (ten_enable_package_manager) {
     deps = [
